Разбивает код на компоненты/классы. Добавляет возможность редактирования каждого ивента
